+++ sources/xcode/Extensions/JLApplicationBadge/JLApplicationBadge/JLApplicationBadge.js
@@ -1,0 +1,13 @@
+(() => {
+    if (window && window.$badge) {
+        return;
+    }
+    
+    const badge = {};
+    badge.set = async (number) => $agent.trigger("$badge.set", number);
+    
+    badge.clear = async () => $agent.trigger("$badge.clear");
+    
+    window.$badge = badge;
+})();
+