+++ sources/xcode/Extensions/JLContacts/JLContacts/JLContacts.m
@@ -67,14 +67,7 @@
     
     // Install the wrappers inside the webview
     
-    NSString * js = [self.app.utils.fs
-                     readJSFor:self];
-    
-    WKUserScript * script = [[WKUserScript alloc] initWithSource:js injectionTime:WKUserScriptInjectionTimeAtDocumentEnd forMainFrameOnly:YES];
-    
-    [webView.configuration.userContentController addUserScript: [script copy]];
-    
-    return webView;
+    return [self.app.utils.webview inject:self intoWebView:webView];
 }